// src/deltas/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 17:53:34 GMT
 */

/**
 * @enum
 * @readonly
 * @description MsigWorlds Table Names
 */
export enum MsigWorldsTableName {
    Approvals = 'approvals',
    Blockedactns = 'blockedactns',
    Invals = 'invals',
    Proposals = 'proposals',
};